+++ core/src/main/scala/at/logic/gapt/proofs/ceres_omega/StandardClauseSet.scala
@@ -5,14 +5,9 @@
 
 package at.logic.gapt.proofs.ceres_omega
 
-import at.logic.gapt.proofs.Sequent
-import at.logic.gapt.proofs.lksk.LKskProof.{ Label, LabelledFormula, LabelledSequent }
+import at.logic.gapt.proofs.{ HOLSequent, Sequent }
 import at.logic.gapt.expr._
-
-import scala.annotation.tailrec
-import scala.util.control.TailCalls._
 import at.logic.gapt.proofs.ceres._
-import at.logic.gapt.utils.Logger
 
 object StandardClauseSet extends StandardClauseSet
 
@@ -27,17 +22,15 @@ class StandardClauseSet {
    * @param contract if contract is set, duplicate formulas will not be added to the clause during a merge (default: true)
    * @return
    */
-  def apply( struct: Struct[Label], contract: Boolean = true ): Set[LabelledSequent] = struct match {
+  def apply( struct: Struct[_], contract: Boolean = true ): Set[HOLSequent] = struct match {
     case A( Top(), _ )    => Set()
     case A( Bottom(), _ ) => Set( Sequent( Nil, Nil ) )
-    case A( fo, label :: Nil ) =>
-      val x: LabelledFormula = ( label, fo )
-      Set( Sequent( Nil, List( x ) ) )
+    case A( fo, _ ) =>
+      Set( Sequent( Nil, List( fo ) ) )
     case Dual( A( Top(), _ ) )    => Set( Sequent( Nil, Nil ) )
     case Dual( A( Bottom(), _ ) ) => Set()
-    case Dual( A( fo, label :: Nil ) ) =>
-      val x: LabelledFormula = ( label, fo )
-      Set( Sequent( List( x ), Nil ) )
+    case Dual( A( fo, _ ) ) =>
+      Set( Sequent( List( fo ), Nil ) )
     case EmptyPlusJunction()            => Set()
     case EmptyTimesJunction()           => Set( Sequent( Nil, Nil ) )
     case Plus( EmptyPlusJunction(), x ) => apply( x )
@@ -64,7 +57,7 @@ class StandardClauseSet {
   }
 
   /* Like compose, but does not duplicate common terms */
-  private def delta_compose( fs1: LabelledSequent, fs2: LabelledSequent ) = Sequent(
+  private def delta_compose( fs1: HOLSequent, fs2: HOLSequent ) = Sequent(
     fs1.antecedent ++ fs2.antecedent.diff( fs1.antecedent ),
     fs1.succedent ++ fs2.succedent.diff( fs1.succedent )
   )