+++ core/src/main/scala/at/logic/gapt/prooftool/TreeVizView.scala
@@ -1,7 +1,6 @@
 package at.logic.gapt.prooftool
 
 import at.logic.gapt.proofs.lk._
-import at.logic.gapt.proofs.lksk
 import at.logic.gapt.proofs.{ DagProof, HOLSequent, SequentProof }
 
 import scala.swing.{ Action, BorderPanel }
@@ -76,19 +75,15 @@ class ProofNodeInfo[T <: DagProof[T]] extends NodeInfo {
   def getColor( path: TreePath2[TreeNode] ) = {
     import Rainbow._
     path.getLastPathComponent.asInstanceOf[ProofNode[T]].proof match {
-      case _: CutRule | _: lksk.Cut | _: Resolution => green
+      case _: CutRule | _: Resolution => green
       case _: InitialSequent | _: WeakeningLeftRule | _: WeakeningRightRule |
         _: ContractionLeftRule | _: ContractionRightRule |
-        _: lksk.InitialSequent | _: lksk.WeakeningLeft | _: lksk.WeakeningRight |
-        _: lksk.ContractionLeft | _: lksk.ContractionRight |
         _: Factor | _: Taut => Color.LIGHT_GRAY
-      case _: AndLeftRule | _: OrRightRule | _: ImpRightRule | _: NegLeftRule | _: NegRightRule |
-        _: lksk.AndLeft | _: lksk.OrRight | _: lksk.ImpRight | _: lksk.NegLeft | _: lksk.NegRight => orange
-      case _: AndRightRule | _: OrLeftRule | _: ImpLeftRule |
-        _: lksk.AndRight | _: lksk.OrLeft | _: lksk.ImpLeft => yellow
-      case WeakQuantifierRule( _, _, _, _, _, _ ) | _: lksk.AllLeft | _: lksk.AllSkLeft | _: lksk.ExRight | _: lksk.ExSkRight => blue
-      case StrongQuantifierRule( _, _, _, _, _ ) | _: lksk.AllRight | _: lksk.AllSkRight | _: lksk.ExLeft | _: lksk.ExSkLeft => red
-      case _: EqualityRule | _: lksk.Equality | _: Paramod | _: Refl | _: Factor => violet
+      case _: AndLeftRule | _: OrRightRule | _: ImpRightRule | _: NegLeftRule | _: NegRightRule => orange
+      case _: AndRightRule | _: OrLeftRule | _: ImpLeftRule => yellow
+      case WeakQuantifierRule( _, _, _, _, _, _ ) => blue
+      case StrongQuantifierRule( _, _, _, _, _ ) | _: SkolemQuantifierRule => red
+      case _: EqualityRule | _: Paramod | _: Refl | _: Factor => violet
       case _: AvatarComponent | _: AvatarContradiction | _: AvatarSplit => yellow
       case _: Subst => orange
       case _ => Color.MAGENTA