+++ tests/src/test/scala/at/logic/gapt/integration_tests/nTapeTest.scala
@@ -2,8 +2,6 @@ package at.logic.gapt.integration_tests
 
 import at.logic.gapt.examples._
 import at.logic.gapt.proofs.Sequent
-import at.logic.gapt.proofs.lksk.LKskProof
-import at.logic.gapt.proofs.lksk.LKskProof.Label
 import at.logic.gapt.provers.eprover.EProver
 import at.logic.gapt.provers.prover9.Prover9
 import org.specs2.mutable._
@@ -12,10 +10,7 @@ class nTapeTest extends Specification {
   args( skipAll = !Prover9.isInstalled )
   "The higher-order tape proof" should {
     "do cut-elimination on the 2 copies tape proof tape3.llk" in {
-      val acnf_labels = nTape2.acnf.conclusion.map( _._1 ).filter( _ != LKskProof.emptyLabel )
-      acnf_labels must_== Sequent[Label]()
-
-      val acnf_lkconclusion = nTape2.acnf.conclusion.map( _._2 ) //discard labels
+      val acnf_lkconclusion = nTape2.acnf.conclusion
       //println( nTape2.preprocessed_input_proof.conclusion )
       //println( acnf_lkconclusion )
       acnf_lkconclusion.multiSetEquals( nTape2.preprocessed_input_proof.conclusion ) must beTrue
@@ -32,10 +27,7 @@ class nTapeTest extends Specification {
     }
 
     "do cut-elimination on the 1 copy tape proof tape3ex.llk" in {
-      val acnf_labels = nTape3.acnf.conclusion.map( _._1 ).filter( _ != LKskProof.emptyLabel )
-      acnf_labels must_== Sequent[Label]()
-
-      val acnf_lkconclusion = nTape3.acnf.conclusion.map( _._2 )
+      val acnf_lkconclusion = nTape3.acnf.conclusion
       acnf_lkconclusion.multiSetEquals( nTape3.preprocessed_input_proof.conclusion ) must beTrue
       nTape3.thf_reproving_deep( None ) must be_!=( "" )